+++ api/http_server_test.go
@@ -1,78 +1,44 @@
-package tests
+package api
 
 import (
 	"fmt"
 	"github.com/golang/protobuf/jsonpb"
-	api "github.com/spacemeshos/go-spacemesh/api"
 	config "github.com/spacemeshos/go-spacemesh/api/config"
 	pb "github.com/spacemeshos/go-spacemesh/api/pb"
 	"github.com/spacemeshos/go-spacemesh/assert"
-	"golang.org/x/net/context"
-	"google.golang.org/grpc"
 	"io/ioutil"
 	"net/http"
-	"strconv"
 	"strings"
 	"testing"
 )
 
-func TestServersConfig(t *testing.T) {
+func TestJsonServerConfig(t *testing.T) {
 
-	config.ConfigValues.GrpcServerPort = 9092
+	// Arrange
 	config.ConfigValues.JsonServerPort = 9031
 
-	grpcService := api.NewGrpcService()
-	jsonService := api.NewJsonHttpServer()
+	// Act
+	jsonService := NewJsonHttpServer()
 
-	assert.Equal(t, grpcService.Port, config.ConfigValues.GrpcServerPort, "Expected same port")
+	// Assert
 	assert.Equal(t, jsonService.Port, config.ConfigValues.JsonServerPort, "Expected same port")
-}
-
-func TestGrpcApi(t *testing.T) {
-
-	const port = 9092
-	const message = "Hello World"
-	config.ConfigValues.GrpcServerPort = port
-
-	grpcService := api.NewGrpcService()
-
-	// start a server
-	grpcService.StartService()
-
-	// start a client
-	addr := "localhost:" + strconv.Itoa(port)
-
-	// Set up a connection to the server.
-	conn, err := grpc.Dial(addr, grpc.WithInsecure())
-	if err != nil {
-		t.Fatalf("did not connect: %v", err)
-	}
-	defer conn.Close()
-	c := pb.NewSpaceMeshServiceClient(conn)
-
-	// call echo and validate result
-	r, err := c.Echo(context.Background(), &pb.SimpleMessage{message})
-	if err != nil {
-		t.Fatalf("could not greet: %v", err)
-	}
-
-	assert.Equal(t, message, r.Value, "Expected message to be echoed")
-
-	// stop the server
-	grpcService.StopService()
 }
 
 func TestJsonApi(t *testing.T) {
 
-	grpcService := api.NewGrpcService()
-	jsonService := api.NewJsonHttpServer()
+	// Arrange
+	const message = "hello world!"
+	const contentType = "application/json"
 
-	// start grp and json server
+	// Act
+	grpcService := NewGrpcService()
+	jsonService := NewJsonHttpServer()
+
 	grpcService.StartService()
 	jsonService.StartService()
 
-	const message = "hello world!"
-	const contentType = "application/json"
+	defer jsonService.Stop()
+	defer grpcService.StopService()
 
 	// generate request payload (api input params)
 	reqParams := pb.SimpleMessage{Value: message}
@@ -115,7 +81,4 @@ func TestJsonApi(t *testing.T) {
 	if value := resp.Header.Get("Content-Type"); value != contentType {
 		t.Errorf("Content-Type was %s, wanted %s", value, contentType)
 	}
-	// stop the services
-	jsonService.Stop()
-	grpcService.StopService()
 }
