+++ app/app_test.go
@@ -1,7 +1,6 @@
-package tests
+package app
 
 import (
-	"github.com/spacemeshos/go-spacemesh/app"
 	"github.com/spacemeshos/go-spacemesh/assert"
 	"os"
 	"testing"
@@ -13,16 +12,16 @@ func TestApp(t *testing.T) {
 	// remove all injected test flags for now
 	os.Args = []string{"/go-spacemesh", "-jrpc"}
 
-	go app.Main("", "master", "")
+	go Main("", "master", "")
 
-	assert.NotNil(t, app.App)
+	assert.NotNil(t, App)
 
 	// let node warmup
 	time.Sleep(3 * time.Second)
 
-	assert.NotNil(t, app.App.Node)
-	assert.NotNil(t, app.App)
+	assert.NotNil(t, App.Node)
+	assert.NotNil(t, App)
 
 	// app should exit based on this signal
-	app.ExitApp <- true
+	ExitApp <- true
 }
